Move PVC related objects to separate package
